requirements.txt
requests
python-dotenv
langchain
unstructured
crewai